+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Changelog
 - Acceptance test for "Sending money" feature should check receiver wallet's balance
 - Improved spending password validation rules
 - Improved acceptance tests for generating new addresses
+- Removed temporary workaround for creating new accounts during wallet create and wallet restore
 - Prevent React key duplicates in transaction from/to addresses lists
 
 ### Chores